First commit
Code in its current state does NOT run correctly. There is an error produced corresponding to a "no decrease in enthalpy in HEM calc". 
Also, OF and chamber temperature are currently not programmed to change with chamber pressure/mass flow rate and are being held at constant. This can easily be changed later but I want the code to at least run first